chore: moves focus indicator from hooks folder to it's own
