+++ change/@fluentui-react-tabster-ad843a71-13a9-4633-862f-9eea9c57a59d.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: moves focus indicator from hooks folder to it's own",
+  "packageName": "@fluentui/react-tabster",
+  "email": "bsunderhus@microsoft.com",
+  "dependentChangeType": "none"
+}
