+++ packages/react-tabster/src/hooks/index.ts
@@ -1,7 +1,6 @@
 export * from './useArrowNavigationGroup';
 export * from './useFocusableGroup';
 export * from './useFocusFinders';
-export * from './useFocusIndicatorStyle';
 export * from './useKeyboardNavAttribute';
 export * from './useModalAttributes';
 export * from './useTabsterAttributes';
